Fix stream destination port register write (4-byte, was 2-byte)
GigE Vision stream channel port register is 4 bytes wide.
Writing only 2 bytes (u16) caused the port value to land in the
wrong position within the register.

+++ crates/tl-gige/src/gvcp.rs
@@ -659,7 +659,7 @@ impl GigeDevice {
         let addr = Self::stream_reg(channel, consts::STREAM_DESTINATION_ADDRESS);
         self.write_mem(addr, &ip.octets()).await?;
         let addr = Self::stream_reg(channel, consts::STREAM_DESTINATION_PORT);
-        self.write_mem(addr, &port.to_be_bytes()).await?;
+        self.write_mem(addr, &(port as u32).to_be_bytes()).await?;
         Ok(())
     }
